+++ client/src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { UserListComponent } from './users/user-list.component';
-import { UserProfileComponent } from './users/user-profile.component';
+import { UserProfileComponent } from './users/user-profile/user-profile.component';
 import { AddUserComponent } from './users/add-user/add-user.component';
 
 // Note that the 'users/new' route needs to come before 'users/:id'.
