+++ client/src/app/app.module.ts
@@ -29,7 +29,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { HomeComponent } from './home/home.component';
 import { UserCardComponent } from './users/user-card/user-card.component';
 import { UserListComponent } from './users/user-list.component';
-import { UserProfileComponent } from './users/user-profile.component';
+import { UserProfileComponent } from './users/user-profile/user-profile.component';
 
 const MATERIAL_MODULES: any[] = [
   MatButtonModule,
